+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/FLTCamSampleBufferTests.m
@@ -8,32 +8,110 @@
 #endif
 @import AVFoundation;
 @import XCTest;
-#import <OCMock/OCMock.h>
 #import "CameraTestUtils.h"
+#import "MockAssetWriter.h"
+#import "MockCaptureConnection.h"
+
+@import camera_avfoundation;
+@import AVFoundation;
+
+@interface FakeMediaSettingsAVWrapper : FLTCamMediaSettingsAVWrapper
+@property(readonly, nonatomic) MockAssetWriterInput *inputMock;
+@end
+
+@implementation FakeMediaSettingsAVWrapper
+- (instancetype)initWithInputMock:(MockAssetWriterInput *)inputMock {
+  _inputMock = inputMock;
+  return self;
+}
+
+- (BOOL)lockDevice:(AVCaptureDevice *)captureDevice error:(NSError **)outError {
+  return YES;
+}
+
+- (void)unlockDevice:(AVCaptureDevice *)captureDevice {
+}
+
+- (void)beginConfigurationForSession:(id<FLTCaptureSession>)videoCaptureSession {
+}
+
+- (void)commitConfigurationForSession:(id<FLTCaptureSession>)videoCaptureSession {
+}
+
+- (void)setMinFrameDuration:(CMTime)duration onDevice:(AVCaptureDevice *)captureDevice {
+}
+
+- (void)setMaxFrameDuration:(CMTime)duration onDevice:(AVCaptureDevice *)captureDevice {
+}
+
+- (id<FLTAssetWriterInput>)assetWriterAudioInputWithOutputSettings:
+    (nullable NSDictionary<NSString *, id> *)outputSettings {
+  return _inputMock;
+}
+
+- (id<FLTAssetWriterInput>)assetWriterVideoInputWithOutputSettings:
+    (nullable NSDictionary<NSString *, id> *)outputSettings {
+  return _inputMock;
+}
+
+- (void)addInput:(AVAssetWriterInput *)writerInput toAssetWriter:(AVAssetWriter *)writer {
+}
+
+- (NSDictionary<NSString *, id> *)
+    recommendedVideoSettingsForAssetWriterWithFileType:(AVFileType)fileType
+                                             forOutput:(AVCaptureVideoDataOutput *)output {
+  return @{};
+}
+@end
 
 /// Includes test cases related to sample buffer handling for FLTCam class.
 @interface FLTCamSampleBufferTests : XCTestCase
-
 @end
 
 @implementation FLTCamSampleBufferTests
 
+- (FLTCam *)createCameraWithAssetWriter:(MockAssetWriter *)assetWriter
+                                adaptor:(MockAssetWriterInputPixelBufferAdaptor *)adaptor
+                                  input:(MockAssetWriterInput *)input {
+  FLTCamConfiguration *configuration = FLTCreateTestCameraConfiguration();
+  configuration.mediaSettings =
+      [FCPPlatformMediaSettings makeWithResolutionPreset:FCPPlatformResolutionPresetMedium
+                                         framesPerSecond:nil
+                                            videoBitrate:nil
+                                            audioBitrate:nil
+                                             enableAudio:YES];
+  configuration.mediaSettingsWrapper = [[FakeMediaSettingsAVWrapper alloc] initWithInputMock:input];
+
+  configuration.assetWriterFactory =
+      ^NSObject<FLTAssetWriter> *_Nonnull(NSURL *url, AVFileType fileType, NSError **error) {
+    return assetWriter;
+  };
+  configuration.inputPixelBufferAdaptorFactory =
+      ^id<FLTAssetWriterInputPixelBufferAdaptor> _Nonnull(NSObject<FLTAssetWriterInput> *input,
+                                                          NSDictionary<NSString *, id> *settings) {
+    return adaptor;
+  };
+
+  return FLTCreateCamWithConfiguration(configuration);
+}
+
 - (void)testSampleBufferCallbackQueueMustBeCaptureSessionQueue {
   dispatch_queue_t captureSessionQueue = dispatch_queue_create("testing", NULL);
-  FLTCam *cam = FLTCreateCamWithCaptureSessionQueue(captureSessionQueue);
-  XCTAssertEqual(captureSessionQueue, cam.captureVideoOutput.sampleBufferCallbackQueue,
+  FLTCam *camera = FLTCreateCamWithCaptureSessionQueue(captureSessionQueue);
+  XCTAssertEqual(captureSessionQueue, camera.captureVideoOutput.sampleBufferCallbackQueue,
                  @"Sample buffer callback queue must be the capture session queue.");
 }
 
 - (void)testCopyPixelBuffer {
-  FLTCam *cam = FLTCreateCamWithCaptureSessionQueue(dispatch_queue_create("test", NULL));
+  FLTCam *camera = FLTCreateCamWithConfiguration(FLTCreateTestCameraConfiguration());
+  MockCaptureConnection *connectionMock = [[MockCaptureConnection alloc] init];
   CMSampleBufferRef capturedSampleBuffer = FLTCreateTestSampleBuffer();
   CVPixelBufferRef capturedPixelBuffer = CMSampleBufferGetImageBuffer(capturedSampleBuffer);
   // Mimic sample buffer callback when captured a new video sample
-  [cam captureOutput:cam.captureVideoOutput
+  [camera captureOutput:camera.captureVideoOutput
       didOutputSampleBuffer:capturedSampleBuffer
-             fromConnection:OCMClassMock([AVCaptureConnection class])];
-  CVPixelBufferRef deliveriedPixelBuffer = [cam copyPixelBuffer];
+             fromConnection:connectionMock];
+  CVPixelBufferRef deliveriedPixelBuffer = [camera copyPixelBuffer];
   XCTAssertEqual(deliveriedPixelBuffer, capturedPixelBuffer,
                  @"FLTCam must deliver the latest captured pixel buffer to copyPixelBuffer API.");
   CFRelease(capturedSampleBuffer);
@@ -41,32 +119,21 @@
 }
 
 - (void)testDidOutputSampleBuffer_mustNotChangeSampleBufferRetainCountAfterPauseResumeRecording {
-  FLTCam *cam = FLTCreateCamWithCaptureSessionQueue(dispatch_queue_create("test", NULL));
+  FLTCam *camera = FLTCreateCamWithConfiguration(FLTCreateTestCameraConfiguration());
+  MockCaptureConnection *connectionMock = [[MockCaptureConnection alloc] init];
   CMSampleBufferRef sampleBuffer = FLTCreateTestSampleBuffer();
 
-  id writerMock = OCMClassMock([AVAssetWriter class]);
-  OCMStub([writerMock alloc]).andReturn(writerMock);
-  OCMStub([writerMock initWithURL:OCMOCK_ANY fileType:OCMOCK_ANY error:[OCMArg setTo:nil]])
-      .andReturn(writerMock);
-  __block AVAssetWriterStatus status = AVAssetWriterStatusUnknown;
-  OCMStub([writerMock startWriting]).andDo(^(NSInvocation *invocation) {
-    status = AVAssetWriterStatusWriting;
-  });
-  OCMStub([writerMock status]).andDo(^(NSInvocation *invocation) {
-    [invocation setReturnValue:&status];
-  });
-
   // Pause then resume the recording.
-  [cam
+  [camera
       startVideoRecordingWithCompletion:^(FlutterError *_Nullable error) {
       }
                   messengerForStreaming:nil];
-  [cam pauseVideoRecording];
-  [cam resumeVideoRecording];
+  [camera pauseVideoRecording];
+  [camera resumeVideoRecording];
 
-  [cam captureOutput:cam.captureVideoOutput
+  [camera captureOutput:camera.captureVideoOutput
       didOutputSampleBuffer:sampleBuffer
-             fromConnection:OCMClassMock([AVCaptureConnection class])];
+             fromConnection:connectionMock];
   XCTAssertEqual(CFGetRetainCount(sampleBuffer), 1,
                  @"didOutputSampleBuffer must not change the sample buffer retain count after "
                  @"pause resume recording.");
@@ -74,55 +141,50 @@
 }
 
 - (void)testDidOutputSampleBufferIgnoreAudioSamplesBeforeVideoSamples {
-  FLTCam *cam = FLTCreateCamWithCaptureSessionQueue(dispatch_queue_create("testing", NULL));
+  MockAssetWriter *writerMock = [[MockAssetWriter alloc] init];
+  MockAssetWriterInputPixelBufferAdaptor *adaptorMock =
+      [[MockAssetWriterInputPixelBufferAdaptor alloc] init];
+  MockAssetWriterInput *inputMock = [[MockAssetWriterInput alloc] init];
+  MockCaptureConnection *connectionMock = [[MockCaptureConnection alloc] init];
+
+  __block AVAssetWriterStatus status = AVAssetWriterStatusUnknown;
+  writerMock.startWritingStub = ^{
+    status = AVAssetWriterStatusWriting;
+  };
+  writerMock.statusStub = ^AVAssetWriterStatus {
+    return status;
+  };
+
+  FLTCam *camera = [self createCameraWithAssetWriter:writerMock
+                                             adaptor:adaptorMock
+                                               input:inputMock];
   CMSampleBufferRef videoSample = FLTCreateTestSampleBuffer();
   CMSampleBufferRef audioSample = FLTCreateTestAudioSampleBuffer();
 
-  id connectionMock = OCMClassMock([AVCaptureConnection class]);
-
-  id writerMock = OCMClassMock([AVAssetWriter class]);
-  OCMStub([writerMock alloc]).andReturn(writerMock);
-  OCMStub([writerMock initWithURL:OCMOCK_ANY fileType:OCMOCK_ANY error:[OCMArg setTo:nil]])
-      .andReturn(writerMock);
-  __block AVAssetWriterStatus status = AVAssetWriterStatusUnknown;
-  OCMStub([writerMock startWriting]).andDo(^(NSInvocation *invocation) {
-    status = AVAssetWriterStatusWriting;
-  });
-  OCMStub([writerMock status]).andDo(^(NSInvocation *invocation) {
-    [invocation setReturnValue:&status];
-  });
-
   __block NSArray *writtenSamples = @[];
 
-  id adaptorMock = OCMClassMock([AVAssetWriterInputPixelBufferAdaptor class]);
-  OCMStub([adaptorMock assetWriterInputPixelBufferAdaptorWithAssetWriterInput:OCMOCK_ANY
-                                                  sourcePixelBufferAttributes:OCMOCK_ANY])
-      .andReturn(adaptorMock);
-  OCMStub([adaptorMock appendPixelBuffer:[OCMArg anyPointer] withPresentationTime:kCMTimeZero])
-      .ignoringNonObjectArgs()
-      .andDo(^(NSInvocation *invocation) {
-        writtenSamples = [writtenSamples arrayByAddingObject:@"video"];
-      });
+  adaptorMock.appendPixelBufferStub = ^BOOL(CVPixelBufferRef buffer, CMTime time) {
+    writtenSamples = [writtenSamples arrayByAddingObject:@"video"];
+    return YES;
+  };
 
-  id inputMock = OCMClassMock([AVAssetWriterInput class]);
-  OCMStub([inputMock assetWriterInputWithMediaType:OCMOCK_ANY outputSettings:OCMOCK_ANY])
-      .andReturn(inputMock);
-  OCMStub([inputMock isReadyForMoreMediaData]).andReturn(YES);
-  OCMStub([inputMock appendSampleBuffer:[OCMArg anyPointer]]).andDo(^(NSInvocation *invocation) {
+  inputMock.readyForMoreMediaData = YES;
+  inputMock.appendSampleBufferStub = ^BOOL(CMSampleBufferRef buffer) {
     writtenSamples = [writtenSamples arrayByAddingObject:@"audio"];
-  });
+    return YES;
+  };
 
-  [cam
+  [camera
       startVideoRecordingWithCompletion:^(FlutterError *_Nullable error) {
       }
                   messengerForStreaming:nil];
 
-  [cam captureOutput:nil didOutputSampleBuffer:audioSample fromConnection:connectionMock];
-  [cam captureOutput:nil didOutputSampleBuffer:audioSample fromConnection:connectionMock];
-  [cam captureOutput:cam.captureVideoOutput
+  [camera captureOutput:nil didOutputSampleBuffer:audioSample fromConnection:connectionMock];
+  [camera captureOutput:nil didOutputSampleBuffer:audioSample fromConnection:connectionMock];
+  [camera captureOutput:camera.captureVideoOutput
       didOutputSampleBuffer:videoSample
              fromConnection:connectionMock];
-  [cam captureOutput:nil didOutputSampleBuffer:audioSample fromConnection:connectionMock];
+  [camera captureOutput:nil didOutputSampleBuffer:audioSample fromConnection:connectionMock];
 
   NSArray *expectedSamples = @[ @"video", @"audio" ];
   XCTAssertEqualObjects(writtenSamples, expectedSamples, @"First appended sample must be video.");
@@ -132,67 +194,58 @@
 }
 
 - (void)testDidOutputSampleBufferSampleTimesMustBeNumericAfterPauseResume {
-  FLTCam *cam = FLTCreateCamWithCaptureSessionQueue(dispatch_queue_create("testing", NULL));
+  MockAssetWriter *writerMock = [[MockAssetWriter alloc] init];
+  MockAssetWriterInputPixelBufferAdaptor *adaptorMock =
+      [[MockAssetWriterInputPixelBufferAdaptor alloc] init];
+  MockAssetWriterInput *inputMock = [[MockAssetWriterInput alloc] init];
+  MockCaptureConnection *connectionMock = [[MockCaptureConnection alloc] init];
+
+  FLTCam *camera = [self createCameraWithAssetWriter:writerMock
+                                             adaptor:adaptorMock
+                                               input:inputMock];
   CMSampleBufferRef videoSample = FLTCreateTestSampleBuffer();
   CMSampleBufferRef audioSample = FLTCreateTestAudioSampleBuffer();
 
-  id connectionMock = OCMClassMock([AVCaptureConnection class]);
-
-  id writerMock = OCMClassMock([AVAssetWriter class]);
-  OCMStub([writerMock alloc]).andReturn(writerMock);
-  OCMStub([writerMock initWithURL:OCMOCK_ANY fileType:OCMOCK_ANY error:[OCMArg setTo:nil]])
-      .andReturn(writerMock);
   __block AVAssetWriterStatus status = AVAssetWriterStatusUnknown;
-  OCMStub([writerMock startWriting]).andDo(^(NSInvocation *invocation) {
+  writerMock.startWritingStub = ^{
     status = AVAssetWriterStatusWriting;
-  });
-  OCMStub([writerMock status]).andDo(^(NSInvocation *invocation) {
-    [invocation setReturnValue:&status];
-  });
+  };
+  writerMock.statusStub = ^AVAssetWriterStatus {
+    return status;
+  };
 
   __block BOOL videoAppended = NO;
-  id adaptorMock = OCMClassMock([AVAssetWriterInputPixelBufferAdaptor class]);
-  OCMStub([adaptorMock assetWriterInputPixelBufferAdaptorWithAssetWriterInput:OCMOCK_ANY
-                                                  sourcePixelBufferAttributes:OCMOCK_ANY])
-      .andReturn(adaptorMock);
-  OCMStub([adaptorMock appendPixelBuffer:[OCMArg anyPointer] withPresentationTime:kCMTimeZero])
-      .ignoringNonObjectArgs()
-      .andDo(^(NSInvocation *invocation) {
-        CMTime presentationTime;
-        [invocation getArgument:&presentationTime atIndex:3];
-        XCTAssert(CMTIME_IS_NUMERIC(presentationTime));
-        videoAppended = YES;
-      });
+  adaptorMock.appendPixelBufferStub = ^BOOL(CVPixelBufferRef buffer, CMTime time) {
+    XCTAssert(CMTIME_IS_NUMERIC(time));
+    videoAppended = YES;
+    return YES;
+  };
 
   __block BOOL audioAppended = NO;
-  id inputMock = OCMClassMock([AVAssetWriterInput class]);
-  OCMStub([inputMock assetWriterInputWithMediaType:OCMOCK_ANY outputSettings:OCMOCK_ANY])
-      .andReturn(inputMock);
-  OCMStub([inputMock isReadyForMoreMediaData]).andReturn(YES);
-  OCMStub([inputMock appendSampleBuffer:[OCMArg anyPointer]]).andDo(^(NSInvocation *invocation) {
-    CMSampleBufferRef sampleBuffer;
-    [invocation getArgument:&sampleBuffer atIndex:2];
-    CMTime sampleTime = CMSampleBufferGetPresentationTimeStamp(sampleBuffer);
+  inputMock.readyForMoreMediaData = YES;
+  inputMock.appendSampleBufferStub = ^BOOL(CMSampleBufferRef buffer) {
+    CMTime sampleTime = CMSampleBufferGetPresentationTimeStamp(buffer);
     XCTAssert(CMTIME_IS_NUMERIC(sampleTime));
     audioAppended = YES;
-  });
+    return YES;
+  };
 
-  [cam
+  [camera
       startVideoRecordingWithCompletion:^(FlutterError *_Nullable error) {
       }
                   messengerForStreaming:nil];
 
-  [cam pauseVideoRecording];
-  [cam resumeVideoRecording];
+  [camera pauseVideoRecording];
+  [camera resumeVideoRecording];
 
-  [cam captureOutput:cam.captureVideoOutput
+  [camera captureOutput:camera.captureVideoOutput
       didOutputSampleBuffer:videoSample
              fromConnection:connectionMock];
-  [cam captureOutput:nil didOutputSampleBuffer:audioSample fromConnection:connectionMock];
-  [cam captureOutput:cam.captureVideoOutput
+  [camera captureOutput:nil didOutputSampleBuffer:audioSample fromConnection:connectionMock];
+  [camera captureOutput:camera.captureVideoOutput
       didOutputSampleBuffer:videoSample
              fromConnection:connectionMock];
-  [cam captureOutput:nil didOutputSampleBuffer:audioSample fromConnection:connectionMock];
+  [camera captureOutput:nil didOutputSampleBuffer:audioSample fromConnection:connectionMock];
   XCTAssert(videoAppended && audioAppended, @"Video or audio was not appended.");
 
   CFRelease(videoSample);
@@ -200,50 +253,38 @@
 }
 
 - (void)testDidOutputSampleBufferMustNotAppendSampleWhenReadyForMoreMediaDataIsNo {
-  FLTCam *cam = FLTCreateCamWithCaptureSessionQueue(dispatch_queue_create("testing", NULL));
+  MockAssetWriter *writerMock = [[MockAssetWriter alloc] init];
+  MockAssetWriterInputPixelBufferAdaptor *adaptorMock =
+      [[MockAssetWriterInputPixelBufferAdaptor alloc] init];
+  MockAssetWriterInput *inputMock = [[MockAssetWriterInput alloc] init];
+  MockCaptureConnection *connectionMock = [[MockCaptureConnection alloc] init];
+  FLTCam *camera = [self createCameraWithAssetWriter:writerMock
+                                             adaptor:adaptorMock
+                                               input:inputMock];
+
   CMSampleBufferRef videoSample = FLTCreateTestSampleBuffer();
 
-  id connectionMock = OCMClassMock([AVCaptureConnection class]);
-
-  id writerMock = OCMClassMock([AVAssetWriter class]);
-  OCMStub([writerMock alloc]).andReturn(writerMock);
-  OCMStub([writerMock initWithURL:OCMOCK_ANY fileType:OCMOCK_ANY error:[OCMArg setTo:nil]])
-      .andReturn(writerMock);
-
   __block BOOL sampleAppended = NO;
-  id adaptorMock = OCMClassMock([AVAssetWriterInputPixelBufferAdaptor class]);
-  OCMStub([adaptorMock assetWriterInputPixelBufferAdaptorWithAssetWriterInput:OCMOCK_ANY
-                                                  sourcePixelBufferAttributes:OCMOCK_ANY])
-      .andReturn(adaptorMock);
-  OCMStub([adaptorMock appendPixelBuffer:[OCMArg anyPointer] withPresentationTime:kCMTimeZero])
-      .ignoringNonObjectArgs()
-      .andDo(^(NSInvocation *invocation) {
-        sampleAppended = YES;
-      });
+  adaptorMock.appendPixelBufferStub = ^BOOL(CVPixelBufferRef buffer, CMTime time) {
+    sampleAppended = YES;
+    return YES;
+  };
 
-  __block BOOL readyForMoreMediaData = NO;
-  id inputMock = OCMClassMock([AVAssetWriterInput class]);
-  OCMStub([inputMock assetWriterInputWithMediaType:OCMOCK_ANY outputSettings:OCMOCK_ANY])
-      .andReturn(inputMock);
-  OCMStub([inputMock isReadyForMoreMediaData]).andDo(^(NSInvocation *invocation) {
-    [invocation setReturnValue:&readyForMoreMediaData];
-  });
-
-  [cam
+  [camera
       startVideoRecordingWithCompletion:^(FlutterError *_Nullable error) {
       }
                   messengerForStreaming:nil];
 
-  readyForMoreMediaData = YES;
+  inputMock.readyForMoreMediaData = YES;
   sampleAppended = NO;
-  [cam captureOutput:cam.captureVideoOutput
+  [camera captureOutput:camera.captureVideoOutput
       didOutputSampleBuffer:videoSample
              fromConnection:connectionMock];
   XCTAssertTrue(sampleAppended, @"Sample was not appended.");
 
-  readyForMoreMediaData = NO;
+  inputMock.readyForMoreMediaData = NO;
   sampleAppended = NO;
-  [cam captureOutput:cam.captureVideoOutput
+  [camera captureOutput:camera.captureVideoOutput
       didOutputSampleBuffer:videoSample
              fromConnection:connectionMock];
   XCTAssertFalse(sampleAppended, @"Sample cannot be appended when readyForMoreMediaData is NO.");
@@ -252,85 +293,80 @@
 }
 
 - (void)testStopVideoRecordingWithCompletionMustCallCompletion {
-  FLTCam *cam = FLTCreateCamWithCaptureSessionQueue(dispatch_queue_create("testing", NULL));
+  MockAssetWriter *writerMock = [[MockAssetWriter alloc] init];
+  MockAssetWriterInputPixelBufferAdaptor *adaptorMock =
+      [[MockAssetWriterInputPixelBufferAdaptor alloc] init];
+  MockAssetWriterInput *inputMock = [[MockAssetWriterInput alloc] init];
+  FLTCam *camera = [self createCameraWithAssetWriter:writerMock
+                                             adaptor:adaptorMock
+                                               input:inputMock];
 
-  id writerMock = OCMClassMock([AVAssetWriter class]);
-  OCMStub([writerMock alloc]).andReturn(writerMock);
-  OCMStub([writerMock initWithURL:OCMOCK_ANY fileType:OCMOCK_ANY error:[OCMArg setTo:nil]])
-      .andReturn(writerMock);
   __block AVAssetWriterStatus status = AVAssetWriterStatusUnknown;
-  OCMStub([writerMock startWriting]).andDo(^(NSInvocation *invocation) {
+  writerMock.startWritingStub = ^{
     status = AVAssetWriterStatusWriting;
-  });
-  OCMStub([writerMock status]).andDo(^(NSInvocation *invocation) {
-    [invocation setReturnValue:&status];
-  });
+  };
+  writerMock.statusStub = ^AVAssetWriterStatus {
+    return status;
+  };
+  writerMock.finishWritingStub = ^(void (^param)(void)) {
+    XCTAssert(writerMock.status == AVAssetWriterStatusWriting,
+              @"Cannot call finishWritingWithCompletionHandler when status is "
+              @"not AVAssetWriterStatusWriting.");
+    void (^handler)(void) = param;
+    handler();
+  };
 
-  OCMStub([writerMock finishWritingWithCompletionHandler:[OCMArg checkWithBlock:^(id param) {
-                        XCTAssert(status == AVAssetWriterStatusWriting,
-                                  @"Cannot call finishWritingWithCompletionHandler when status is "
-                                  @"not AVAssetWriterStatusWriting.");
-                        void (^handler)(void) = param;
-                        handler();
-                        return YES;
-                      }]]);
-
-  [cam
+  [camera
       startVideoRecordingWithCompletion:^(FlutterError *_Nullable error) {
       }
                   messengerForStreaming:nil];
 
   __block BOOL completionCalled = NO;
-  [cam stopVideoRecordingWithCompletion:^(NSString *_Nullable path, FlutterError *_Nullable error) {
-    completionCalled = YES;
-  }];
+  [camera
+      stopVideoRecordingWithCompletion:^(NSString *_Nullable path, FlutterError *_Nullable error) {
+        completionCalled = YES;
+      }];
   XCTAssert(completionCalled, @"Completion was not called.");
 }
 
 - (void)testStartWritingShouldNotBeCalledBetweenSampleCreationAndAppending {
-  FLTCam *cam = FLTCreateCamWithCaptureSessionQueue(dispatch_queue_create("testing", NULL));
+  MockAssetWriter *writerMock = [[MockAssetWriter alloc] init];
+  MockAssetWriterInputPixelBufferAdaptor *adaptorMock =
+      [[MockAssetWriterInputPixelBufferAdaptor alloc] init];
+  MockAssetWriterInput *inputMock = [[MockAssetWriterInput alloc] init];
+  MockCaptureConnection *connectionMock = [[MockCaptureConnection alloc] init];
+  FLTCam *camera = [self createCameraWithAssetWriter:writerMock
+                                             adaptor:adaptorMock
+                                               input:inputMock];
+
   CMSampleBufferRef videoSample = FLTCreateTestSampleBuffer();
 
-  id connectionMock = OCMClassMock([AVCaptureConnection class]);
-
-  id writerMock = OCMClassMock([AVAssetWriter class]);
-  OCMStub([writerMock alloc]).andReturn(writerMock);
-  OCMStub([writerMock initWithURL:OCMOCK_ANY fileType:OCMOCK_ANY error:[OCMArg setTo:nil]])
-      .andReturn(writerMock);
   __block BOOL startWritingCalled = NO;
-  OCMStub([writerMock startWriting]).andDo(^(NSInvocation *invocation) {
+  writerMock.startWritingStub = ^{
     startWritingCalled = YES;
-  });
+  };
 
   __block BOOL videoAppended = NO;
-  id adaptorMock = OCMClassMock([AVAssetWriterInputPixelBufferAdaptor class]);
-  OCMStub([adaptorMock assetWriterInputPixelBufferAdaptorWithAssetWriterInput:OCMOCK_ANY
-                                                  sourcePixelBufferAttributes:OCMOCK_ANY])
-      .andReturn(adaptorMock);
-  OCMStub([adaptorMock appendPixelBuffer:[OCMArg anyPointer] withPresentationTime:kCMTimeZero])
-      .ignoringNonObjectArgs()
-      .andDo(^(NSInvocation *invocation) {
-        videoAppended = YES;
-      });
+  adaptorMock.appendPixelBufferStub = ^BOOL(CVPixelBufferRef buffer, CMTime time) {
+    videoAppended = YES;
+    return YES;
+  };
 
-  id inputMock = OCMClassMock([AVAssetWriterInput class]);
-  OCMStub([inputMock assetWriterInputWithMediaType:OCMOCK_ANY outputSettings:OCMOCK_ANY])
-      .andReturn(inputMock);
-  OCMStub([inputMock isReadyForMoreMediaData]).andReturn(YES);
+  inputMock.readyForMoreMediaData = YES;
 
-  [cam
+  [camera
       startVideoRecordingWithCompletion:^(FlutterError *_Nullable error) {
       }
                   messengerForStreaming:nil];
 
   BOOL startWritingCalledBefore = startWritingCalled;
-  [cam captureOutput:cam.captureVideoOutput
+  [camera captureOutput:camera.captureVideoOutput
       didOutputSampleBuffer:videoSample
              fromConnection:connectionMock];
   XCTAssert((startWritingCalledBefore && videoAppended) || (startWritingCalled && !videoAppended),
             @"The startWriting was called between sample creation and appending.");
 
-  [cam captureOutput:cam.captureVideoOutput
+  [camera captureOutput:camera.captureVideoOutput
       didOutputSampleBuffer:videoSample
              fromConnection:connectionMock];
   XCTAssert(videoAppended, @"Video was not appended.");
@@ -341,15 +377,9 @@
 - (void)testStartVideoRecordingWithCompletionShouldNotDisableMixWithOthers {
   FLTCam *cam = FLTCreateCamWithCaptureSessionQueue(dispatch_queue_create("testing", NULL));
 
-  id writerMock = OCMClassMock([AVAssetWriter class]);
-  OCMStub([writerMock alloc]).andReturn(writerMock);
-  OCMStub([writerMock initWithURL:OCMOCK_ANY fileType:OCMOCK_ANY error:[OCMArg setTo:nil]])
-      .andReturn(writerMock);
-
   [AVAudioSession.sharedInstance setCategory:AVAudioSessionCategoryPlayback
                                  withOptions:AVAudioSessionCategoryOptionMixWithOthers
                                        error:nil];
-
   [cam
       startVideoRecordingWithCompletion:^(FlutterError *_Nullable error) {
       }