+++ packages/video_player/video_player_avfoundation/darwin/video_player_avfoundation/Sources/video_player_avfoundation_ios/FVPNativeVideoView.m
@@ -1,0 +1,38 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#import "../video_player_avfoundation/include/video_player_avfoundation/FVPNativeVideoView.h"
+
+#import <AVFoundation/AVFoundation.h>
+
+@interface FVPPlayerView : UIView
+@end
+
+@implementation FVPPlayerView
++ (Class)layerClass {
+  return [AVPlayerLayer class];
+}
+
+- (void)setPlayer:(AVPlayer *)player {
+  [(AVPlayerLayer *)[self layer] setPlayer:player];
+}
+@end
+
+@interface FVPNativeVideoView ()
+@property(nonatomic) FVPPlayerView *playerView;
+@end
+
+@implementation FVPNativeVideoView
+- (instancetype)initWithPlayer:(AVPlayer *)player {
+  if (self = [super init]) {
+    _playerView = [[FVPPlayerView alloc] init];
+    [_playerView setPlayer:player];
+  }
+  return self;
+}
+
+- (FVPPlayerView *)view {
+  return self.playerView;
+}
+@end